First take on proper connection handling
Changes include:
 - thread separation from GUI and connection stuff
  - connection estabilishing structure tweaks
 - byte proccessing fix
 - testing server addr + port configuration
 - command type returning fix (command toString as a bonus)
 - a lot of network exceptions are now secured as well

+++ src/generic/SimpleScanner.java
@@ -39,8 +39,8 @@ public class SimpleScanner {
 		return source.length() - pointer;
 	}
 
-	public byte nextByte() {
-		return Byte.parseByte(read(2), parse_radix);
+	public short nextByte() {
+		return Short.parseShort(read(2), parse_radix);
 	}
 
 	public short nextShort() {
